requirements.txt
tkinter-0.1.0
qrcode-7.3.1
screeninfo-0.8
pan-aadhar-ocr